+++ src/main/java/searchengine/repositories/PageRepository.java
@@ -15,13 +15,13 @@ public interface PageRepository extends JpaRepository<PageEntity,Integer> {
     PageEntity findByName(String path);
 
     @Query(value = " select * from page where `path` =:path and site_id =:siteId", nativeQuery = true)
-    PageEntity findByPathAndId(String path, int siteId);
+    PageEntity findByPathAndId(String path, long siteId);
 
     @Query(value = "select * from page where site_id=:siteId", nativeQuery = true)
-    ArrayList<PageEntity> findAllPagesForSite(int siteId);
+    ArrayList<PageEntity> findAllPagesForSite(long siteId);
 
     @Query(value = "select count(*) from page where site_id =:siteId", nativeQuery = true)
-    int countPagesForSite(int siteId);
+    int countPagesForSite(long siteId);
 
 
 }
